+++ packages/common/SupportedHostings.ts
@@ -1,0 +1,6 @@
+export type SupportedHostings = 'github'
+    | 'githubgist'
+    | 'gitlab'
+    | 'bitbucket'
+    | 'pastebin'
+    | 'sourceforge';