+++ packages/popup/Popup.tsx
@@ -2,9 +2,11 @@ import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { LocalStorage } from '../common/LocalStorage';
 import { sendMessage } from '../common/Messenger';
+import { getHostData } from '../common/HostData';
 
 type State = {
-    storage: LocalStorage
+    storage: LocalStorage,
+    isSomethingChanged: boolean
 };
 
 type Props = {
@@ -15,7 +17,8 @@ class Popup extends React.Component<Props, State> {
     constructor(props: Props) {
         super(props);
         this.state = {
-            storage: props.storage
+            storage: props.storage,
+            isSomethingChanged: false
         };
     }
 
@@ -30,7 +33,8 @@ class Popup extends React.Component<Props, State> {
         };
         sendMessage({ type: 'STORAGE_SET', storage: newStorage });
         this.setState({
-            storage: newStorage
+            storage: newStorage,
+            isSomethingChanged: true
         });
     }
 
@@ -43,24 +47,31 @@ class Popup extends React.Component<Props, State> {
 
     render() {
         const hostings = Object.keys(this.props.storage.showIcons) as (keyof LocalStorage['showIcons'])[];
+        const changedText = this.state.isSomethingChanged
+            ? (<p style={{ color: 'orange' }}><i>To see changes, please reload pages, their states you changed</i></p>)
+            : null;
         return (
             <div id="settings">
-                <h3>vscode-icons for:</h3>
+                <h3>display icons for:</h3>
                 <div>
-                    {hostings.map((hosting, index) => (
-                        <div key={index} className="form-group" style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
-                            <label className="form-checkbox">
-                                {hosting}
-                                <input
-                                    type="checkbox"
-                                    checked={this.state.storage.showIcons[hosting]}
-                                    onChange={(_: any) => this.handleToggleClick(hosting)}
-                                />
-                                <i className="form-icon" />
-                            </label>
-                            <img src={chrome.runtime.getURL(`favicons/${hosting}-favicon.ico`)} width="16" height="16" />
-                        </div>
-                    ))}
+                    {hostings.map((hosting, index) => {
+                        const hostData = getHostData(hosting);
+                        return (
+                            <div key={index} className="form-group" style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
+                                <label className="form-checkbox">
+                                    {hostData.fullName}
+                                    <input
+                                        type="checkbox"
+                                        checked={this.state.storage.showIcons[hosting]}
+                                        onChange={(_: any) => this.handleToggleClick(hosting)}
+                                    />
+                                    <i className="form-icon" />
+                                </label>
+                                <img className="vsi-icon" src={chrome.runtime.getURL(`favicons/${hostData.favicon}`)} />
+                            </div>
+                        );
+                    })}
+                    {changedText}
                 </div>
                 {/* <button onClick={this.handleResetButton}>Reset</button> */}
             </div>