+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
-## [1.0.7] - 2017-11-DD
+## [1.0.7] - 2017-12-DD
 
 ### Added
 
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
     - Display icons in user profile
 - Support for [sourceforge.com](https://sourceforge.net)
     - Display icons in project files
+- Added Popup, where you can turn off/on vsi for specific hostings
 
 ### Changed